chg: move checkProcessByPid and sleep to helpers
