+++ src/lib/wpr/WprAbstract.ts
@@ -1,31 +1,14 @@
+import {ChildProcess, spawn} from 'child_process';
 import * as fs from 'fs';
 
 import {Logger} from '@/lib/logger';
 import {IWprConfig, IWprProcess} from '@/lib/wpr/types';
-import {ChildProcess, spawn} from 'child_process';
+import {checkProcessByPid, sleep} from './helpers';
 
 export type BuildCmd = (wprConfig: IWprConfig, inputWprFilepath: string) => {command: string, args: string[]};
 
 export const WAIT_TIMEOUT = 10000;
 export const SLEEP_TIMEOUT_BEFORE_CHECK_PROCESS = 500;
-
-export function checkProcessByPid(pid: number) {
-    try {
-        return process.kill(pid, 0);
-    } catch (error) {
-        if (error.code === 'EPERM' || error.code === 'ESRCH') {
-            return false;
-        } else {
-            throw error;
-        }
-    }
-}
-
-export function sleep(ms: number) {
-    return new Promise((resolve) => {
-        setTimeout(() => { resolve(); }, ms);
-    });
-}
 
 export default abstract class WprAbstract implements IWprProcess {
     protected _logger: Logger;
